The startups form component was created
